Fix qHash(NormalizedRect)

+++ core/area.h
@@ -967,7 +967,7 @@ public:
     QColor color;
 };
 
-uint qHash(const Okular::NormalizedRect &r, uint seed = 0);
+size_t qHash(const Okular::NormalizedRect &r, size_t seed = 0);
 }
 
 #ifndef QT_NO_DEBUG_STREAM
